Refactor exhaustivity checker to be testable
